% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-example_data.R
\docType{data}
\name{airline_tweets}
\alias{airline_tweets}
\title{Airline Tweet Data}
\format{
An object of class \code{data.table} with the "text" column required for
text processing and the "sentiment" column for checking positive
or negative predictions of the model.
}
\usage{
data(airline_tweets)
}
\description{
This is a basic data that can be used to test out the installation and make
sure that the model is running correctly. The current data is based on airline
tweets with corresponding sentiment labels.
}
\keyword{datasets}
\keyword{examples,}
