% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_data.R
\docType{data}
\name{default_embeddings}
\alias{default_embeddings}
\title{Default sentiment embeddings}
\format{
An object of class `"list"`
}
\usage{
data(default_embeddings)
}
\description{
This is a default sentiment embeddings for the default data. You can use them
for vector ordination; otherwise they exist to speed up sentiment analysis
when you have positive = NULL, negative = NULL, and are using either a default
"en" or "multi" model.
}
\keyword{datasets}
