% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_and_install.R
\name{.activate_env}
\alias{.activate_env}
\title{Activate sentiment.ai environment in reticulate}
\usage{
.activate_env(envname = "r-sentiment-ai", silent = FALSE, r_envir = -1)
}
\description{
Activate sentiment.ai environment in reticulate
}
