% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sentiment.R
\name{.reference_embed}
\alias{.reference_embed}
\title{Needs to return a list of embeddings AND the category/sentiment of each row}
\usage{
.reference_embed(positive = NULL, negative = NULL, model)
}
\description{
Needs to return a list of embeddings AND the category/sentiment of each row
}
