% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sentiment.R
\name{.sentiment_probs}
\alias{.sentiment_probs}
\title{Apply model to embeddings to get sentiment score}
\usage{
.sentiment_probs(embeddings, scoring, scoring_version, model)
}
\description{
Apply model to embeddings to get sentiment score
}
