% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sentiment.R
\name{find_sentiment_probs}
\alias{find_sentiment_probs}
\title{Apply Model for Sentiment Score}
\usage{
find_sentiment_probs(embeddings, scoring, scoring_version, model)
}
\description{
Apply Model for Sentiment Score
}
