% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sentiment.R
\name{get_defualt_embedding}
\alias{get_defualt_embedding}
\title{get default embedding
If it exists, return the object
If not, try downloading it
If download works, return object
Else return NULL (to be handles in embed_topics())}
\usage{
get_defualt_embedding(model)
}
\arguments{
\item{model}{character - whic model's default embeddings are needed}
}
\description{
get default embedding
If it exists, return the object
If not, try downloading it
If download works, return object
Else return NULL (to be handles in embed_topics())
}
