% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/embedding.R
\name{openai_embed}
\alias{openai_embed}
\title{Embed text using OpenAI}
\usage{
openai_embed(text, request_limit = 3000, token_limit = 3e+05)
}
\arguments{
\item{text}{A character vector of text to embed.}

\item{request_limit}{The maximum number of requests allowed.}

\item{token_limit}{The maximum number of tokens allowed.}
}
\value{
A data.table containing the embeddings.
}
\description{
Embed text using OpenAI
}
\examples{
\dontrun{
  init_sentiment.ai(...)
  result <- openai_embed(c("hello", "world"))
}
}
