% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sentiment.R
\name{read_embedding}
\alias{read_embedding}
\title{Read embedding file.
Take json path, return single embedding object for specific model.}
\usage{
read_embedding(file, model = "en.large", version = NULL)
}
\arguments{
\item{file}{character - the filepath to the json object.}

\item{model}{character - which model's default embeddings are needed.}

\item{version}{PLACEHOLDER - may be necessary in future.}
}
\description{
Read embedding file.
Take json path, return single embedding object for specific model.
}
