% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object-sentiment_ai_embed.R
\docType{data}
\name{sentiment.ai_embed}
\alias{sentiment.ai_embed}
\title{Sentiment AI Embedding Environment}
\format{
An object of class \code{environment} of length 0.
}
\usage{
sentiment.ai_embed
}
\description{
This is the embedding envirnoment for the sentiment.ai model. On package load,
this object should be NULL. When the model is initialized, this should be
updated based on the required embedding model. The embedding function will
be stored in the "f" slot of this environment.
}
\keyword{datasets}
