% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sentiment.R
\name{sentiment_model}
\alias{sentiment_model}
\title{Model-Based Sentiment}
\usage{
sentiment_model(
  x = NULL,
  data = NULL,
  idcol = NULL,
  lexicon = NULL,
  model = names(default_models),
  ...
)
}
\arguments{
\item{x}{A plain text vector or column name if data is supplied}

\item{data}{A data.frame or data.table with text}

\item{idcol}{If data is supplied, use this column to pick something?}

\item{lexicon}{A data.frame or data.table of words: sentiment (default is XXX)}

\item{model}{An embedding name from tensorflow-hub, some of which are
"en" (english large or not) and "multi" (multi-lingual large or not).}

\item{...}{Additional arguments passed to \code{\link[reticulate:conda-tools]{conda_install()}}
or \code{\link[reticulate:virtualenv-tools]{virtualenv_install()}}.}
}
\description{

}
