% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sentiment.R
\name{sentiment_simple}
\alias{sentiment_simple}
\title{Simple Sentiment}
\usage{
sentiment_simple(
  x = NULL,
  model = names(default_models),
  scoring = c("xgb", "glm"),
  scoring_version = "1.0",
  batch_size = 100,
  ...
)
}
\arguments{
\item{x}{A plain text vector or column name if data is supplied}

\item{model}{An embedding name from tensorflow-hub, some of which are
"en" (english large or not) and "multi" (multi-lingual large or not).}

\item{scoring}{Model to use for scoring the embedding matrix (currently
either "xgb" or "glm").}

\item{batch_size}{Size of batches to use. Larger numbers will be faster than
smaller numbers, but do not exhaust your system memory!}

\item{...}{Additional arguments passed to \code{\link[reticulate:conda-tools]{conda_install()}}
or \code{\link[reticulate:virtualenv-tools]{virtualenv_install()}}.}
}
\description{
This uses a simple model (xgboost or glm) to return a simple predictive score,
where numbers closer to 1 are more positive and numbers closer to -1 are more
negative. This can be used to determine whether the sentiment is positive
or negative.
}
\details{
Uses simple preditive models on embeddings to provide probability of positive
score.
}
\examples{

}
