% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sentiment.R
\name{as_py_list}
\alias{as_py_list}
\title{as py list
because R to Python conversion doesn't work with list is of length 1}
\usage{
as_py_list(x)
}
\arguments{
\item{x}{character vector that is to be passed into tensorflowtext via reticulate.}
}
\value{
List if x is length 1 else x.
}
\description{
as py list
because R to Python conversion doesn't work with list is of length 1
}
