% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_data.R
\docType{data}
\name{default}
\alias{default}
\title{Default sentiment matching dictionary}
\format{
An object of class `"data.table"`
}
\usage{
data(default)
}
\description{
This is a default sentiment matching dictionary with over 100 pairs of positive:negative examples
Feel free to use as-is or use as an example to create your own.
The main point is that there needs to be a corresponding negative for each positive.
For example:
pos <- c("good apples", "fresh", "delicious")
neg <- c("bad apples", "not fresh", "not delicious")
}
\details{
if, for example, positive was:
c("good", "fresh", "delicious")

then

"These were some good apples"

would be seen as closer to a negative example!
}
\examples{
data(default)
View(default)
}
\keyword{datasets}
