% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sentiment.R
\name{embed_text}
\alias{embed_text}
\title{Create Text Embedding Matrix}
\usage{
embed_text(text, batch_size = NULL, model = NULL)
}
\arguments{
\item{text}{character vector to be embedded. Notye that longer comments take longer}

\item{batch_size}{integer - how many to embed at once. Higher numbers are faster but use more memory.}

\item{model}{character - the embedding model to use (same as sentiment_score())}
}
\description{
turns charactewr vector into length(text)x512 embedding matrix.
For power users. Requires init_sentiment.ai() to have been called!
}
