% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_and_install.R
\name{install_default_embeddings}
\alias{install_default_embeddings}
\title{Function to grab the default embeddings for sentiment_match()
Necessary to keep package size under 5Mb.
Will check if they're there, if so return TRUE
if not try download and return TRUE
else return FALSE (and generate them - will take a few seconds!)}
\usage{
install_default_embeddings()
}
\description{
Function to grab the default embeddings for sentiment_match()
Necessary to keep package size under 5Mb.
Will check if they're there, if so return TRUE
if not try download and return TRUE
else return FALSE (and generate them - will take a few seconds!)
}
