% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_helpers.R
\name{matrix_similarity}
\alias{matrix_similarity}
\alias{cosine}
\alias{rescale}
\title{Cosine Similarity}
\usage{
cosine(x, y = NULL)

rescale(x)
}
\arguments{
\item{x}{A numeric vector or matrix}

\item{y}{A numeric vector or matrix of the same dimensions as x}
}
\description{
Cosine Similarity
}
\examples{
\dontrun{
n <- 1000
y <- matrix(rnorm(n * 512), ncol = 512)
x <- matrix(rnorm(n * 512), ncol = 512)

microbenchmark::microbenchmark(
  cosine(x, y),
  text2vec::sim2(x, y),
  times = 10
)

all.equal(cosine(x, y),
          text2vec::sim2(x, y))
}

}
