% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sentiment.R
\name{read_embedding}
\alias{read_embedding}
\title{read embedding file
take json path, return single embedding object for specific model}
\usage{
read_embedding(file, model = "en.large", version = NULL)
}
\arguments{
\item{file}{character - the filepath to the json object}

\item{model}{character - whic model's default embeddings are needed}

\item{version}{PLACEHOLDER - may be necessary in future.}
}
\description{
read embedding file
take json path, return single embedding object for specific model
}
