% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-sentiment_ai.R
\docType{package}
\name{sentiment.ai-package}
\alias{sentiment.ai}
\alias{sentiment.ai-package}
\title{sentiment.ai: Simplified sentiment analysis with AI}
\description{
This package uses Google's Universal Sentence Encoder, simplifies all the
difficulties, and turns it into a sentiment analysis package.

Main Benefits:
\itemize{
  \item{Tolerates spelling mistakes}
  \item{Not dependent on exactly matching a fixed dictionary}
  \item{Requires less pre-processing}
  \item{More powerful than dictionary-based methods}
}

Main Drawbacks:
\itemize{
  \item{Requires a lot of RAM}
  \item{Can be slow on larger datasets (unless using GPU)}
}

Effectivelly, if you have a reasonably powerful computer, you can use
sentiment.ai as a more flexible, powerful, and modern approach to sentiment
analysis.
}
\author{
\strong{Maintainer}: Ben Wiseman \email{benjamin.h.wiseman@gmail.com}

Authors:
\itemize{
  \item Steven Nydick \email{swnydick@gmail.com} (\href{https://orcid.org/0000-0002-2908-1188}{ORCID})
  \item Tristan Wisner \email{tristan.wisner@kornferry.com}
}

Other contributors:
\itemize{
  \item Fiona Lodge \email{fiona.lodge@kornferry.com} [contributor]
  \item Yu-Ann Wang \email{yu-ann-wang@kornferry.com} [contributor]
}

}
