% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sentiment.R
\name{sentiment_easy}
\alias{sentiment_easy}
\title{Simpler Sentiment - give score ranging from -1 (negative) to +1 (positive)
This returns a single vector, and will run a little faster than sentiment_match or sentiment_model
Defaults to English-specific, change model to 'multi' for support of 16 languages from Universal Sentence Encoder Multilingual.}
\usage{
sentiment_easy(
  x = NULL,
  model = c("en.large", "multi.large", "en", "multi"),
  scoring = c("xgb", "glm"),
  scoring_version = "1.0",
  batch_size = 100,
  envname = "r-sentiment-ai"
)
}
\arguments{
\item{x}{a plain text vector or column name if data supplied}

\item{model}{embedding from tensorflow-hub. Shortcuts are en (english-large) and multi (multi-lingual-large)}

\item{scoring}{model used to score embedding. Can choose between a glm or xbg.}

\item{scoring_version}{Placeholder - future versions may include updates/improvements}

\item{batch_size}{Compute embeddings in n batches. The higher the faster BUT be careful not to exhaust your system memory!}

\item{envname}{specify virtual environment for Reticulate}
}
\description{
Simpler Sentiment - give score ranging from -1 (negative) to +1 (positive)
This returns a single vector, and will run a little faster than sentiment_match or sentiment_model
Defaults to English-specific, change model to 'multi' for support of 16 languages from Universal Sentence Encoder Multilingual.
}
