% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sentiment.R
\name{sentiment_match}
\alias{sentiment_match}
\alias{sentiment_model}
\title{Sentiment Matching - give score AND explanation
This returns a single vector, and will run a little faster}
\usage{
sentiment_match(
  x = NULL,
  model = c("en.large", "multi.large", "en", "multi"),
  positive = default$positive,
  negative = default$negative,
  scoring = c("xgb", "glm"),
  scoring_version = "1.0",
  batch_size = 100,
  envname = "r-sentiment-ai"
)

sentiment_model(
  x = NULL,
  data = NULL,
  idcol = NULL,
  lexicon = NULL,
  model = "multi",
  envname = "r-sentiment-ai"
)
}
\arguments{
\item{x}{EITHER a plain text vector or column name if data supplied}

\item{model}{embedding from tensorflow-hub}

\item{positive}{Custom positive word or term to compare against. e.g. "happy", "high quality"}

\item{negative}{Custom Negative word or term to compare against. e.g. "unhappy", "low quality"}

\item{envname}{specify virtual environment for Reticulate}

\item{data}{OPTIONAL dataframe or data.table with text}

\item{idcol}{OPTIONAL IF data supplied, use idcol to do shit}

\item{lexicon}{data.frame or data.table of words: sentiment (default is XXX)}
}
\value{
data.table with sentiment score, topic match, and match score
}
\description{
Sentiment Matching - give score AND explanation
This returns a single vector, and will run a little faster
}
