% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_and_install.R
\name{install_sentiment.ai}
\alias{install_sentiment.ai}
\alias{sentiment.ai.init}
\title{Sets up environment specific for sentiment.as
Currently needs:}
\usage{
install_sentiment.ai(
  envname = "r-sentiment-ai",
  method = c("auto", "virtualenv", "conda"),
  gpu = FALSE,
  conda = "auto",
  python_version = "3.7.10",
  modules = list(numpy = "1.19.5", sentencepiece = "0.1.95", tensorflow = "2.4.1",
    tensorflow_hub = "0.12.0", `tensorflow-text` = "2.4.3"),
  restart_session = TRUE,
  ...
)

sentiment.ai.init(
  model = c("en.large", "multi.large", "en", "multi"),
  envname = "r-sentiment-ai"
)
}
\arguments{
\item{method}{Installation method. By default, "auto" automatically finds a
method that will work in the local environment. Change the default to force
a specific installation method. Note that the "virtualenv" method may not
available on Windows (issue in TensorFlow). Note also
that since this command runs without privilege the "system" method is
available only on Windows.}

\item{gpu}{logical, should GPU enabled TensorFlow be installed?}

\item{model}{path to tensorflow hub embedding model.default are universal sentence encoder  en (default) and multi}
}
\description{
| Module          | Version |
| :-------------- | :------:|
|python           | 3.7.10  |
|numpy            | 1.19.5  |
|tensorflow       | 2.4.1   |
|tensorflow_hub   | 0.12.0  |
|tensorflow-text  | 2.4.3   |
|sentencepiece    | 0.1.95  |
}
\details{
Installs with like tensorflow::install_tensorflow with pip=TRUE
}
